+++ packages/router-plugin/tests/code-splitter/snapshots/react/3-all-combined-errorComponent-separate/lowercase-function-declaration@component---loader---notFoundComponent---pendingComponent.tsx
@@ -1,0 +1,8 @@
+function component() {
+  return <div>lowercase function declaration</div>;
+}
+function pendingComponent() {
+  return <div>lowercase pending function declaration</div>;
+}
+export { component };
+export { pendingComponent };